requirements.txt
autogen-agentchat
autogen-ext[openai,groq]

langgraph
langchain
langchain-core
langchain-community

groq
openai

guardrails-ai
nemoguardrails

tavily-python
requests
beautifulsoup4
aiohttp

semanticscholar

streamlit
gradio
flask
fastapi
uvicorn

python-dotenv
pydantic
pyyaml

pytest
black

# Pre-commit hooks for security and code quality
pre-commit
detect-secrets
bandit
